+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## Release 0.1.0
+## Release 0.1.1
 
 **Features**
 
+Now use the bultin type for the system defaults too for more coverage
+
 **Bugfixes**
 
+Correctly initialize gssproxy::services
+
 **Known Issues**
+
+## Release 0.1.0